+++ src/path_stats.ts
@@ -1,4 +1,8 @@
 import { sep, dirname } from 'path';
 
 /** Gets _output_ directory name. */
-export const getBuildDirectory = (url: string) => dirname(url).split(sep)[0];
+export const getBuildDirectory = (url: string) => {
+  const directoryName = dirname(url).split(sep)[0];
+
+  return directoryName === '.' ? '' : directoryName;
+};